Move database and service logic to seperate files
