+++ api/src/models.py
@@ -1,0 +1,19 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class HttpError:
+	""" https://blog.restcase.com/rest-api-error-handling-problem-details-response """
+
+	""" A human-readable description of the specific error. """
+	detail: str
+
+	""" a URL to a document describing the error condition """
+	type: str
+
+	""" A short, human-readable title for the general error type """
+	title: str
+
+	""" HTTP status code """
+	status: int
+
